Don't cache handle data.
HandleManager used to cache some of the data. However, it has no
knowledge of database rollbacks so when a rollback occurred, its
cached data would diverge from what was in the database.

This cache was removed to prevent this problem. The alternative would
have been to detect rollbacks, which is overly complicated.