Change how empty elements are handled.
When wed uses ``innerHTML`` to pass the edited data to the server,
empty elements have an opening and closing tag. e.g. ``<foo></foo>``
rather than ``<foo/>``. The lxml package, however, wants to output the
shorter, single tag version. This change makes it so that it does the
same thing as ``innerHTML``.