Introduced drop_entry_lock.
release_entry_lock would fail hard if an entry was unlocked when there
is no lock on it or if the lock belongs to another user. However, as
part of normal operations this can happen if the lock was broken by an
administrator or if the lock expired and another user obtained a lock on
the article. drop_entry_lock will just do nothing if the entry is not
locked or if another user acquired the lock.