Modified bibliography search.
Previously, the filtering of bibliographical entries would perform a
match on secondary sources and primary sources independently. So it was
possible to have a match on a secondary source and have none of its
associated primary sources match. Showing the primary sources of such a
secondary source, after filtering, would show no primary source. It
turns out this does not mesh well with the way the authors work, so the
search is now changed so that if a secondary source matches, then all of
its primary sources are also considered to match.