Use execute_async_script more carefully.
Calls to the callback that execute_async_script provide to the code it
executes were done carelessly. In some cases, it seemed possible that
the same callback could be called more than once. Unfortunately,
Selenium does not protect against this. Consider:

- execute_async_script("something")
- execute_async_script("something else")

If the first call is such that it calls its callback once but leaves
behind something that could call the callback again, then there's a risk
that the second execute_async_script will be terminated due to code from
the first call calling its own callback!