Upgrade to behave 1.2.5.
Behave 1.2.5 requires that ``print`` be in the function form rather than
the statement form. Moreover the regular expression engine now forces ^
and $ at the end of the patterns and so raises an error if the patterns
already have these.

The upgrade is necessary as BTW's testing code is about to use a feature
that only exists in behave >= 1.2.5.