Revoke scheduled tasks on start.
btwworker was modified to revoke scheduled tasks when the workers are
started. Celery's list of scheduled tasks persists from run to
run. Therefore the bibliography worker in particular was increasing the
number of scheduled tasks by one each time it was started. This change
ensure old tasks are revoked rather than run.