Tooltip fix for Bootstrap 3.3.5 and over.
The default trigger for tooltips is a combination of ``focus`` and
``hover``. The earlier versions of Bootstrap had a bug that made it so
that the combination did not work correctly. Our code was inadvertently
depending on this bug. 3.3.5 fixed the Bootstrap bug, which changed the
behavior that we were depending on and thus caused problems. The issue
has been fixed.