Do not show edit button if editing is impossible.
There was a bug which caused the edit button to be shown even if editing
was not possible. This was because ``edit_url`` was set to ``None`` in
Python code, which became the string "None" in JavaScript. We make
``edit_url`` be the empty string if editing is impossible, which fixes
the issue.