Use the schema_version field for sorting.
The previous code would use the schema_version property that appears on
ChangeRecord. This made it impossible to actually sort on schema
version. We instead use the schema_version on the associated Chunk.

Also, django-datatables-view uses the dot notation to access related
fields, rather than the double underscore notation proper to Django. So
we fixed this too.