Work around Django CMS's problems.
Django CMS records the "current user" with a ``set_current_user`` function
which can have repercussions in testing. The recent upgrades made it so
the that problem with ``set_current_user`` spread all over the place. So
we patch the appropriate Django test classes to make them clear the
current user rather than have to change all tests everywhere.

This change also flags a new method on ``BaseCMSTestCase`` as "not a
test". Otherwise nose tries to run it.