Perform ad-hoc cleaning between tests.
We no longer use the @dirty decorator but instead systematically perform
ad-hoc cleaning between tests. The problem with letting Django doing the
cleanup was that it could result in situations that are not meant to
happen in production, like for instance Chunk objects being
deleted. Modifying the code of the application to handle these *special
testing cases* was introducing complications that would exist only for
testing and would also probably mask problems that could be detected in
testing.