Always call the overriden method.
Previous implementation would sometimes short-circuit and omit calling
the overriden method. This meant that the test code was not waiting for
the subprocess to end, which caused a bunch of child processes to remain
behind for a while. Updated code always calls the overriden function but
returns an fake response when needed.