Let runserver serve static files (don't use --nostatic).
At some point in the past we used --nostatic to prevent runserver from
intercepting requests to static files and bypassing the entire
middleware. Unfortunately, we did not note why we did that. On the
other hand, it causes problems. Since each request for static files
pass through the entire middleware stack, they may hit the database
and use more connections than what Postgres allows. So we returned to
the default behavior.