Upgrade mitmproxy
We had kept mitmproxy pinned to a very old version. It started causing problems
basically because it did not respond properly to TLSv1.3 requests (which seems
to now be the default). We're still on Python 2 but it requires Python 3. At any
rate, we managed to work around the Python 3 issues. Side benefits: we've been
able to eliminate a few rigmaroles to work around bugs in the old version.