Upgrade to Django CMS 3.6.
We remove django-reversion. It was used only by Django CMS and is not used by
anything else. (It is supposed to be replaced with something else so we do not
remove the tables, in case, there's eventually an upgrade path.)

The test code has been reworked to make the creation of cms pages a bit slicker.

final/apps.py had to be updated to include a new admin view provided by Django
CMS.