Get rid of filer, and filer-related CMS plugins.
They've become a liability. The filer-related CMS plugins that we used are no
longer maintained and will go boom when Python 3.8 becomes current. They've been
replaced by other plugins but there's no trivial upgrade path to move the data
from the old plugins to the new ones (Thanks Divio!). The best thing to do for
us is to just dump the old plugins and fix the CMS pages manually. There
probably won't be much to fix because the assistants that eventually took over
the editing of the CMS pages mostly did away with using the filer facilities. (I
think there's only one remaining instance of an image managed by filer.)