Repatriate most configuration options into the stock settings file.
Over the history of the project, more and more options have been added. It was
not always clear whether these options would be ones that change from site to
site (dev vs prouction vs demo) or not. It turns out that a good deal of the
options are stable irrespective of where BTW is deployed. So we've repatriated
these options into the stock settings file, which makes the site-specific
setting files much smaller and manageable.