test: increate time delays
We increase the time delays because it seems BrowserStack has slowed down in
some fashion.

Loading marionette-glue was timing out on the first test so we had to raise
waitSeconds.

The spinner tests were failing... probably because BrowserStack slowed down the
tests such that the 1 second delay was not long enough. By the time the test was
looking for the spinner, the spinner was gone.