test: actually create a chunk
It is a mistery as to why the tests that depend on the fixed view ever
passed. The view was simply not saving the new data to the database.

The tests failing coincide with the move to eXist-db 5. Was there a bug in
eXist-db's prevous versions that cancelled the bug here??